Code quality fixes and refactoring
