+++ src/main/java/com/example/apgw/controller/UserInfoController.java
@@ -1,6 +1,6 @@
 package com.example.apgw.controller;
 
-import com.example.apgw.model.User;
+import com.example.apgw.helper.User;
 import com.example.apgw.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
